+++ src/command/convert/pdf-to-text/index.ts
@@ -2,25 +2,34 @@ import { isEmpty } from 'lodash';
 import { execSync } from 'child_process';
 import * as fs from 'fs';
 import { getDocumentData, getDocumentFilePath } from '../../../data';
+import { DocumentNode } from '../../../legal/document';
 
-export function pdf2txt(): void {
-  const legals = getDocumentData('pdf');
-  legals.forEach((legal) => {
-    const pdfPath = getDocumentFilePath(legal, 'pdf');
-    const textPath = getDocumentFilePath(legal, 'text');
+type Option = { overwrite: boolean };
+export function pdfToTxt(option: Option): void {
+  const pdfs = getDocumentData('pdf');
+  pdfs.forEach((pdf) => handlePdf(pdf, option));
+}
 
-    try {
-      PyMuPDF(pdfPath, textPath);
-
-      const rawText = fs.readFileSync(textPath).toString();
-      const processedText = postProcess(rawText);
-      fs.writeFileSync(textPath, processedText);
-
-      console.log(`Finished pdf2text ${textPath}`);
-    } catch {
-      console.log(`Error pdf2text ${textPath}`);
+function handlePdf(pdf: DocumentNode, option: Option): void {
+  const { overwrite } = option;
+  const pdfFile = getDocumentFilePath(pdf, 'pdf');
+  const { path: textPath, exists: textExists } = getDocumentFilePath(pdf, 'text');
+  try {
+    if (!overwrite && textExists) {
+      console.log(`Skipped pdf-to-text ${textPath}`);
+      return;
     }
-  });
+
+    PyMuPDF(pdfFile.path, textPath);
+
+    const rawText = fs.readFileSync(textPath).toString();
+    const processedText = postProcess(rawText);
+    fs.writeFileSync(textPath, processedText);
+
+    console.log(`Finished pdf-to-text ${textPath}`);
+  } catch {
+    console.log(`Error pdf-to-text ${textPath}`);
+  }
 }
 
 function postProcess(raw: string): string {
@@ -32,5 +41,5 @@ function postProcess(raw: string): string {
 }
 
 function PyMuPDF(pdfPath: string, textPath: string): void {
-  execSync(`/usr/bin/python3 script/pdf2text/PyMuPDF/index.py ${pdfPath} ${textPath}`);
+  execSync(`/usr/bin/python3 script/pdf-to-text/PyMuPDF/index.py ${pdfPath} ${textPath}`);
 }
